+++ packages/core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sap/guided-answers-extension-core
 
+## 0.1.0
+
+### Minor Changes
+
+-   cc3d9f7: Adding functionality to call Guided Answers command with node path
+
+### Patch Changes
+
+-   Updated dependencies [cc3d9f7]
+    -   @sap/guided-answers-extension-types@0.3.0
+
 ## 0.0.3
 
 ### Patch Changes